# docs/swagger.yaml
definitions:
  domain.Buyer:
    properties:
      card_number_id:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
    type: object
  domain.Carrie:
    properties:
      address:
        type: string
      cid:
        type: string
      company_name:
        type: string
      id:
        type: integer
      locality_id:
        type: string
      telephone:
        type: string
    required:
    - address
    - cid
    - company_name
    - locality_id
    - telephone
    type: object
  domain.CarrieLocality:
    properties:
      carries_count:
        type: integer
      local_name:
        type: string
      locality_id:
        type: string
    required:
    - carries_count
    - local_name
    - locality_id
    type: object
...
  web.errorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  web.response:
    properties:
      data: {}
    type: object
info:
  contact: {}
  description: This API was created as part of Mercado Libre's Go backend bootcamp
    for training purposes.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es-ar/terminos-y-condiciones
  title: MeLi Bootcamp API
  version: "1.0"
paths:
  /api/v1/buyers:
    get:
      description: Returns a list of all buyers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Buyer'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: List buyers
      tags:
      - Buyers
    post:
      consumes:
      - application/json
      description: Create buyer
      parameters:
      - description: buyers parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Buyer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Buyer'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Create buyer
      tags:
      - Buyers
  /api/v1/buyers/{id}:
    delete:
      description: Delete buyer
      parameters:
      - description: buyer id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Delete buyer
      tags:
      - Buyers
    get:
      description: get buyer by id
      parameters:
      - description: buyer id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Buyer'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Buyer by id
      tags:
      - Buyers
    patch:
      consumes:
      - application/json
      description: Update buyer
      parameters:
      - description: buyer id
        in: path
        name: id
        required: true
        type: integer
      - description: buyer parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Buyer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Buyer'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Update buyer
      tags:
      - Buyers
  /api/v1/buyers/reportPurchaseOrders:
    get:
      description: get report by id or all buyers
      parameters:
      - description: Buyer ID
        in: query
        name: id
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Buyer'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Purchase orders by buyer and all
      tags:
      - Buyers
  /api/v1/carries:
    post:
      consumes:
      - application/json
      description: Create carry
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Carrie'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Carrie'
              type: object
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Create carry
      tags:
      - Carry
...
  /api/v1/localities/reportCarries:
    get:
      description: Returns a list of all carries by locality
      parameters:
      - description: locality Id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.CarrieLocality'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: count carries by locality
      tags:
      - Carry
...
securityDefinitions:
  ApiKeyAuth:
    description: start with Bearer
    in: header
    name: x-tiger-token
    type: apiKey
swagger: "2.0"